requirements.txt
altair                    5.0.1
annotated-types           0.6.0
anyio                     4.2.0
argon2-cffi               21.3.0
argon2-cffi-bindings      21.2.0
asttokens                 2.0.5
async-lru                 2.0.4
attrs                     23.1.0
Babel                     2.11.0
beautifulsoup4            4.12.2
bleach                    4.1.0
blinker                   1.6.2
Bottleneck                1.3.7
branca                    0.7.1
Brotli                    1.0.9
cachetools                4.2.2
certifi                   2024.2.2
cffi                      1.16.0
charset-normalizer        2.0.4
click                     8.1.7
click-plugins             1.1.1
cligj                     0.7.2
colorama                  0.4.6
coloredlogs               15.0.1
comm                      0.2.1
contourpy                 1.2.1
cycler                    0.12.1
debugpy                   1.6.7
decorator                 5.1.1
defusedxml                0.7.1
distro                    1.9.0
exceptiongroup            1.2.0
executing                 0.8.3
fastjsonschema            2.16.2
fiona                     1.9.6
flatbuffers               24.3.25
folium                    0.16.0
fonttools                 4.51.0
geographiclib             2.0
geopandas                 0.14.3
geopy                     2.4.1
gitdb                     4.0.7
GitPython                 3.1.37
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.0
humanfriendly             10.0
idna                      3.4
imageio                   2.34.0
ipykernel                 6.28.0
ipython                   8.20.0
ipython-genutils          0.2.0
ipywidgets                7.6.5
IUCN-API                  0.1.6
jedi                      0.18.1
Jinja2                    3.1.3
joblib                    1.4.0
json5                     0.9.6
jsonschema                4.19.2
jsonschema-specifications 2023.7.1
jupyter_client            8.6.0
jupyter_core              5.5.0
jupyter-events            0.8.0
jupyter-lsp               2.2.0
jupyter_server            2.10.0
jupyter_server_terminals  0.4.4
jupyterlab                4.0.11
jupyterlab-pygments       0.1.2
jupyterlab_server         2.25.1
jupyterlab-widgets        3.0.10
kiwisolver                1.4.5
lazy_loader               0.4
llvmlite                  0.42.0
markdown-it-py            2.2.0
MarkupSafe                2.1.3
matplotlib                3.8.4
matplotlib-inline         0.1.6
mdurl                     0.1.0
mistune                   2.0.4
mkl-fft                   1.3.8
mkl-random                1.2.4
mkl-service               2.4.0
mpmath                    1.3.0
nbclient                  0.8.0
nbconvert                 7.10.0
nbformat                  5.9.2
nest-asyncio              1.6.0
networkx                  3.3
notebook                  7.0.8
notebook_shim             0.2.3
numba                     0.59.1
numexpr                   2.8.7
numpy                     1.26.4
onnxruntime               1.17.3
openai                    1.17.0
opencv-python             4.9.0.80
opencv-python-headless    4.9.0.80
overrides                 7.4.0
packaging                 23.2
pandas                    2.2.1
pandocfilters             1.5.0
parso                     0.8.3
pillow                    10.2.0
pip                       23.3.1
platformdirs              3.10.0
plotly                    5.21.0
pooch                     1.8.1
prometheus-client         0.14.1
prompt-toolkit            3.0.43
protobuf                  3.20.3
psutil                    5.9.0
pure-eval                 0.2.2
pyarrow                   14.0.2
pycparser                 2.21
pydantic                  2.7.0
pydantic_core             2.18.1
pydeck                    0.8.0
Pygments                  2.15.1
PyMatting                 1.1.12
pyparsing                 3.1.2
pyproj                    3.6.1
pyreadline3               3.4.1
PySocks                   1.7.1
python-dateutil           2.8.2
python-dotenv             1.0.1
python-json-logger        2.0.7
pytz                      2023.3.post1
pywin32                   305.1
pywinpty                  2.0.10
PyYAML                    6.0.1
pyzmq                     25.1.2
referencing               0.30.2
rembg                     2.0.56
requests                  2.31.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rich                      13.3.5
rpds-py                   0.10.6
scikit-image              0.23.1
scikit-learn              1.4.2
scipy                     1.13.0
seaborn                   0.13.2
Send2Trash                1.8.2
setuptools                68.2.2
shapely                   2.0.4
six                       1.16.0
smmap                     4.0.0
sniffio                   1.3.0
soupsieve                 2.5
stack-data                0.2.0
streamlit                 1.32.0
streamlit_folium          0.19.0
sympy                     1.12
tenacity                  8.2.2
terminado                 0.17.1
threadpoolctl             3.4.0
tifffile                  2024.2.12
tinycss2                  1.2.1
toml                      0.10.2
tomli                     2.0.1
toolz                     0.12.0
tornado                   6.3.3
tqdm                      4.66.2
traitlets                 5.7.1
typing_extensions         4.9.0
tzdata                    2023.3
urllib3                   2.1.0
watchdog                  2.1.6
wcwidth                   0.2.5
webencodings              0.5.1
websocket-client          0.58.0
wheel                     0.41.2
widgetsnbextension        3.5.2
win-inet-pton             1.1.0
xyzservices               2024.4.0